backend/requirements.txt
Flask==3.0.3
flask_cors==4.0.1
psycopg2-binary==2.9.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.1
python-dotenv==1.0.0